--- a/groups_test.go
+++ b/groups_test.go
@@ -58,3 +58,11 @@ func TestAppendAttrsToGroup(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkUniqAttrs(b *testing.B) {
+	var actualAttrs = []slog.Attr{slog.String("key1", "value1"), slog.String("key1", "value1"), slog.String("key2", "value2")}
+	for i := 0; i < b.N; i++ {
+		unique := UniqAttrs(actualAttrs)
+		_ = unique
+	}
+}