```
goos: linux
goarch: amd64
pkg: github.com/samber/slog-common
cpu: AMD Ryzen 9 5950X 16-Core Processor            
             │ baseline.txt │               pr.txt                │
             │    sec/op    │   sec/op     vs base                │
UniqAttrs-32    77.78n ± 1%   12.37n ± 1%  -84.10% (p=0.000 n=10)

             │ baseline.txt │              pr.txt               │
             │     B/op     │   B/op    vs base                 │
UniqAttrs-32     128.0 ± 0%   0.0 ± 0%  -100.00% (p=0.000 n=10)

             │ baseline.txt │               pr.txt                │
             │  allocs/op   │ allocs/op   vs base                 │
UniqAttrs-32     1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
```